experimental_testing/api_auto_xxx/requirements.txt
aniso8601==4.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==2.1.7
djangorestframework==3.9.1
Flask==1.0.2
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
meaninglessTestPackage==0.1.0
pkg-resources==0.0.0
pycodestyle==2.5.0
python-decouple==3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
